+++ api/constants.ts
@@ -1,4 +1,3 @@
 export const ROOT_PATH = "/home/app/standalone";
-export const BUILD_PATH = "/home/app/build";
 export const TIDAL_API_URL = "https://api.tidal.com";
 export const SYNC_DEFAULT_CRON = "0 3 * * *";