+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Tidarr notable changes.
 ### 🚀 Added
 * [Front] Add custom CSS #383
 * [API] Add umask env var for output file chmod #396
+* [API] Refacto server sent events (performances should be significantly enhanced) 
 
 ## 📦 0.3.8
 ### 🖍 Changed